backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "celery>=5.5.3",
    "coverage>=7.13.2",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.61.0",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "redis>=6.4.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.32",
]
